dashboard file

+++ src/app/login/login.component.ts
@@ -126,15 +126,15 @@ export class LoginComponent implements OnInit {
       this.generate()
       return;
     }
-    // if (this.loginForm.controls.passwordForm.value['passwordCaptcha'] != this.captcha ) {
-    //   this.loginForm.patchValue({
-    //     passwordForm: {
-    //       passwordCaptcha: '',
-    //     }
-    //   })
-    //   this.generate()
-    //   return;
-    // }
+    if (this.loginForm.controls.passwordForm.value['passwordCaptcha'] != this.captcha ) {
+      this.loginForm.patchValue({
+        passwordForm: {
+          passwordCaptcha: '',
+        }
+      })
+      this.generate()
+      return;
+    }
       sessionStorage.clear();
       sessionStorage.removeItem('authTeacherDetails')
       sessionStorage.removeItem('mappingData')
@@ -182,15 +182,15 @@ export class LoginComponent implements OnInit {
       this.generate();
       return;
     }
-    // if (this.loginForm.controls.otpForm.value['otpCaptcha'] != this.captchaotp ) {
-    //   this.loginForm.patchValue({
-    //     otpForm: {
-    //       otpCaptcha: '',
-    //     }
-    //   })
-    //   this.generate();
-    //   return;
-    // }
+    if (this.loginForm.controls.otpForm.value['otpCaptcha'] != this.captchaotp ) {
+      this.loginForm.patchValue({
+        otpForm: {
+          otpCaptcha: '',
+        }
+      })
+      this.generate();
+      return;
+    }
     sessionStorage.clear();
     sessionStorage.removeItem('authTeacherDetails')
     sessionStorage.removeItem('mappingData')
